--- Acme-Jobs/src/main/java/acme/features/auditor/job/AuditorJobListMineService.java
@@ -1,48 +1,0 @@
-
-package acme.features.auditor.job;
-
-import java.util.Collection;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import acme.entities.jobs.Job;
-import acme.entities.roles.Auditor;
-import acme.framework.components.Model;
-import acme.framework.components.Request;
-import acme.framework.entities.Principal;
-import acme.framework.services.AbstractListService;
-
-@Service
-public class AuditorJobListMineService implements AbstractListService<Auditor, Job> {
-
-	@Autowired
-	AuditorJobRepository repository;
-
-
-	@Override
-	public boolean authorise(final Request<Job> request) {
-		assert request != null;
-
-		return true;
-	}
-	@Override
-	public void unbind(final Request<Job> request, final Job entity, final Model model) {
-		assert request != null;
-		assert entity != null;
-		assert model != null;
-
-		request.unbind(entity, model, "reference", "title", "deadline");
-	}
-	@Override
-	public Collection<Job> findMany(final Request<Job> request) {
-		assert request != null;
-		Collection<Job> result;
-		Principal principal;
-
-		principal = request.getPrincipal();
-		result = this.repository.findManyByEmployerId(principal.getActiveRoleId());
-
-		return result;
-	}
-}
