--- Acme-Jobs/src/main/java/acme/features/auditor/job/AuditorJobShowService.java
@@ -1,47 +1,0 @@
-
-package acme.features.auditor.job;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import acme.entities.jobs.Job;
-import acme.entities.roles.Auditor;
-import acme.framework.components.Model;
-import acme.framework.components.Request;
-import acme.framework.services.AbstractShowService;
-
-@Service
-public class AuditorJobShowService implements AbstractShowService<Auditor, Job> {
-
-	@Autowired
-	AuditorJobRepository repository;
-
-
-	@Override
-	public boolean authorise(final Request<Job> request) {
-		assert request != null;
-
-		return true;
-	}
-	@Override
-	public void unbind(final Request<Job> request, final Job entity, final Model model) {
-		assert request != null;
-		assert entity != null;
-		assert model != null;
-
-		request.unbind(entity, model, "reference", "title", "deadline");
-		request.unbind(entity, model, "salary", "moreInfo", "description", "draft");
-	}
-	@Override
-	public Job findOne(final Request<Job> request) {
-		assert request != null;
-
-		Job result;
-		int id;
-
-		id = request.getModel().getInteger("id");
-		result = this.repository.findOneJobById(id);
-
-		return result;
-	}
-}
