--- Acme-Jobs/src/main/java/acme/features/auditor/writtenJob/AuditorJobController.java
@@ -1,29 +1,0 @@
-
-package acme.features.auditor.writtenJob;
-
-import javax.annotation.PostConstruct;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import acme.components.CustomCommand;
-import acme.entities.jobs.Job;
-import acme.entities.roles.Auditor;
-import acme.framework.components.BasicCommand;
-import acme.framework.controllers.AbstractController;
-
-@Controller
-@RequestMapping("/auditor/job-written/")
-public class AuditorJobController extends AbstractController<Auditor, Job> {
-
-	@Autowired
-	private AuditorJobListMineService listService;
-
-
-	@PostConstruct
-	private void initialise() {
-		super.addCustomCommand(CustomCommand.LIST_MINE, BasicCommand.LIST, this.listService);
-	}
-
-}
