+++ Acme-Jobs/src/main/java/acme/features/auditor/writtenJob/AuditorWrittenJobRepository.java
@@ -1,5 +1,5 @@
 
-package acme.features.auditor.job;
+package acme.features.auditor.writtenJob;
 
 import java.util.Collection;
 
@@ -10,8 +10,12 @@ import acme.entities.jobs.Job;
 import acme.framework.repositories.AbstractRepository;
 
 @Repository
-public interface AuditorJobRepository extends AbstractRepository {
+public interface AuditorWrittenJobRepository extends AbstractRepository {
 
-	@Query("select j from Job j join AuditRecord ar on ar.job.id=j.id where ar.auditor.id=?1")
+	@Query("select j from Job j where j.id=?1")
+	Job findOneJobById(int id);
+
+	@Query("select j from Job j join AuditRecord ar on ar.job.id=j.id where ar.auditor.id!=?1")
 	Collection<Job> findManyByEmployerId(int auditorId);
+
 }
