+++ Acme-Jobs/src/main/java/acme/features/auditor/writtenJob/AuditorWrittenJobShowService.java
@@ -1,8 +1,6 @@
 
 package acme.features.auditor.writtenJob;
 
-import java.util.Collection;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,14 +8,13 @@ import acme.entities.jobs.Job;
 import acme.entities.roles.Auditor;
 import acme.framework.components.Model;
 import acme.framework.components.Request;
-import acme.framework.entities.Principal;
-import acme.framework.services.AbstractListService;
+import acme.framework.services.AbstractShowService;
 
 @Service
-public class AuditorJobListMineService implements AbstractListService<Auditor, Job> {
+public class AuditorWrittenJobShowService implements AbstractShowService<Auditor, Job> {
 
 	@Autowired
-	AuditorJobRepository repository;
+	AuditorWrittenJobRepository repository;
 
 
 	@Override
@@ -33,15 +30,17 @@ public class AuditorJobListMineService implements AbstractListService<Auditor, J
 		assert model != null;
 
 		request.unbind(entity, model, "reference", "title", "deadline");
+		request.unbind(entity, model, "salary", "moreInfo", "description", "draft");
 	}
 	@Override
-	public Collection<Job> findMany(final Request<Job> request) {
+	public Job findOne(final Request<Job> request) {
 		assert request != null;
-		Collection<Job> result;
-		Principal principal;
 
-		principal = request.getPrincipal();
-		result = this.repository.findManyByEmployerId(principal.getActiveRoleId());
+		Job result;
+		int id;
+
+		id = request.getModel().getInteger("id");
+		result = this.repository.findOneJobById(id);
 
 		return result;
 	}
