+++ Acme-Jobs/src/main/webapp/META-INF/views/master/i18n_es.messages
@@ -94,11 +94,8 @@ master.menu.employer.application.list-mine = Listar aplicaciones
 
 master.menu.auditor = Auditor
 master.menu.auditor.features.audit-record.list = Listar registros de auditoría
-master.menu.auditor.features.job.list = Listar mis trabajos
+master.menu.auditor.features.job.list = Listar trabajos que contienen mis registros de auditoría
 
-# master.menu.auditor ##########################################################
-master.menu.auditor = Auditor
-master.menu.auditor.features.job.list = Listar mis trabajos
 # master.menu.user-account ####################################################
 
 master.menu.user-account = Cuenta
